+++ packages/pandas-gbq/tests/system.py
@@ -50,12 +50,8 @@ def _get_dataset_prefix_random():
 
 
 def _get_project_id():
-
-    project = os.environ.get('GBQ_PROJECT_ID')
-    if not project:
-        pytest.skip(
-            "Cannot run integration tests without a project id")
-    return project
+    return (os.environ.get('GBQ_PROJECT_ID')
+            or os.environ.get('GOOGLE_CLOUD_PROJECT'))  # noqa
 
 
 def _get_private_key_path():
@@ -85,9 +81,12 @@ def _test_imports():
     gbq._test_google_api_imports()
 
 
-@pytest.fixture
-def project():
-    return _get_project_id()
+@pytest.fixture(params=['env'])
+def project(request):
+    if request.param == 'env':
+        return _get_project_id()
+    elif request.param == 'none':
+        return None
 
 
 def _check_if_can_get_correct_default_credentials():
@@ -99,11 +98,13 @@ def _check_if_can_get_correct_default_credentials():
     from google.auth.exceptions import DefaultCredentialsError
 
     try:
-        credentials, _ = google.auth.default(scopes=[gbq.GbqConnector.scope])
+        credentials, project = google.auth.default(
+            scopes=[gbq.GbqConnector.scope])
     except (DefaultCredentialsError, IOError):
         return False
 
-    return gbq._try_credentials(_get_project_id(), credentials) is not None
+    return gbq._try_credentials(
+        project or _get_project_id(), credentials) is not None
 
 
 def clean_gbq_environment(dataset_prefix, private_key=None):
@@ -171,46 +172,14 @@ def test_generate_bq_schema_deprecated():
         gbq.generate_bq_schema(df)
 
 
-@pytest.fixture(params=['local', 'service_path', 'service_creds'])
-def auth_type(request):
-
-    auth = request.param
-
-    if auth == 'local':
-
-        if _in_travis_environment():
-            pytest.skip("Cannot run local auth in travis environment")
-
-    elif auth == 'service_path':
-
-        if _in_travis_environment():
-            pytest.skip("Only run one auth type in Travis to save time")
-
-        _skip_if_no_private_key_path()
-    elif auth == 'service_creds':
-        _skip_if_no_private_key_contents()
-    else:
-        raise ValueError
-    return auth
-
-
 @pytest.fixture()
-def credentials(auth_type):
-
-    if auth_type == 'local':
-        return None
-
-    elif auth_type == 'service_path':
-        return _get_private_key_path()
-    elif auth_type == 'service_creds':
-        return _get_private_key_contents()
-    else:
-        raise ValueError
+def credentials():
+    _skip_if_no_private_key_contents()
+    return _get_private_key_contents()
 
 
 @pytest.fixture()
 def gbq_connector(project, credentials):
-
     return gbq.GbqConnector(project, private_key=credentials)
 
 
@@ -220,7 +189,7 @@ class TestGBQConnectorIntegration(object):
         assert gbq_connector is not None, 'Could not create a GbqConnector'
 
     def test_should_be_able_to_get_valid_credentials(self, gbq_connector):
-        credentials = gbq_connector.get_credentials()
+        credentials, _ = gbq_connector.get_credentials()
         assert credentials.valid
 
     def test_should_be_able_to_get_a_bigquery_client(self, gbq_connector):
@@ -236,14 +205,12 @@ class TestGBQConnectorIntegration(object):
         assert pages is not None
 
 
-class TestGBQConnectorIntegrationWithLocalUserAccountAuth(object):
+class TestAuth(object):
 
     @pytest.fixture(autouse=True)
-    def setup(self, project):
-
-        _skip_local_auth_if_in_travis_env()
-
-        self.sut = gbq.GbqConnector(project, auth_local_webserver=True)
+    def setup(self, gbq_connector):
+        self.sut = gbq_connector
+        self.sut.auth_local_webserver = True
 
     def test_get_application_default_credentials_does_not_throw_error(self):
         if _check_if_can_get_correct_default_credentials():
@@ -252,9 +219,9 @@ class TestGBQConnectorIntegrationWithLocalUserAccountAuth(object):
             from google.auth.exceptions import DefaultCredentialsError
             with mock.patch('google.auth.default',
                             side_effect=DefaultCredentialsError()):
-                credentials = self.sut.get_application_default_credentials()
+                credentials, _ = self.sut.get_application_default_credentials()
         else:
-            credentials = self.sut.get_application_default_credentials()
+            credentials, _ = self.sut.get_application_default_credentials()
         assert credentials is None
 
     def test_get_application_default_credentials_returns_credentials(self):
@@ -262,10 +229,14 @@ class TestGBQConnectorIntegrationWithLocalUserAccountAuth(object):
             pytest.skip("Cannot get default_credentials "
                         "from the environment!")
         from google.auth.credentials import Credentials
-        credentials = self.sut.get_application_default_credentials()
+        credentials, default_project = (
+            self.sut.get_application_default_credentials())
+
         assert isinstance(credentials, Credentials)
+        assert default_project is not None
 
     def test_get_user_account_credentials_bad_file_returns_credentials(self):
+        _skip_local_auth_if_in_travis_env()
 
         from google.auth.credentials import Credentials
         with mock.patch('__main__.open', side_effect=IOError()):
@@ -273,6 +244,8 @@ class TestGBQConnectorIntegrationWithLocalUserAccountAuth(object):
         assert isinstance(credentials, Credentials)
 
     def test_get_user_account_credentials_returns_credentials(self):
+        _skip_local_auth_if_in_travis_env()
+
         from google.auth.credentials import Credentials
         credentials = self.sut.get_user_account_credentials()
         assert isinstance(credentials, Credentials)
@@ -515,7 +488,8 @@ class TestReadGBQIntegration(object):
 
     def test_bad_project_id(self):
         with pytest.raises(gbq.GenericGBQException):
-            gbq.read_gbq("SELECT 1", project_id='001',
+            gbq.read_gbq('SELCET * FROM [publicdata:samples.shakespeare]',
+                         project_id='not-my-project',
                          private_key=self.credentials)
 
     def test_bad_table_name(self):